ft(Avail request): Avail requests for approval - Get all requests - Add search functionality - Add pagination
[Finishes #170764534]
