+++ src/components/ServerPagination/ServerSidePagination.scss
@@ -1,73 +1,81 @@
-
-.rest_pagination_items{
-  height: auto;
+.pagination-container {
   display: flex;
   flex-direction: row;
-  justify-content: space-between;
-  width: 70%;
-  margin: 0 auto;
-  float: none;
+  justify-content: center;
+  padding: 1.5em 1em;
+  font-size: 1em;
+  font-family: 'Poppins', sans-serif;
+  color: #555 !important;
+  background: #ffffff;
+}
 
-  &_dropdown {
-    width: 80px;
-    margin-left: 40px;
-    border: 1px solid rgba(196, 196, 196, 0.58);
+.pagination-section {
+  display: flex;
+  align-items: center;
+  justify-content: center;
 
+  &__text {
+    margin: 0 1.25em;
+
+    span {
+      color: #555 !important;
+    }
+  }
+ 
+  &__select {
+    width: 75px;
+    display: inline-block;
+    padding: 2px 6px 2px 0;
+    box-sizing: border-box;
+    line-height: 1.42857143;
+    background-color: #ffffff;
+    background-image: none;
+    border: none;
+    outline: none;
+    -webkit-box-shadow: inset 0 -2px rgba(0, 0, 0, 0.075);
+    box-shadow: inset 0 -2px rgba(0, 0, 0, 0.075);
+    -webkit-transition: border-color ease-in-out 0.15s, -webkit-box-shadow ease-in-out 0.15s;
+    -o-transition: border-color ease-in-out 0.15s, box-shadow ease-in-out 0.15s;
+    transition: border-color ease-in-out 0.15s, box-shadow ease-in-out 0.15s;
   }
 
-  &_currentPage{
-      margin-left: 25px;
-      padding-top: 15px;
-  }
-  
-  &_rowsPerPage{
-      margin-left: 25px;
-      padding: 20px;
-      padding-top: 15px;
-  }
-  
-  &_allPages{
-      margin-left: 25px;
-      padding-top: 15px;
-  }
-  
-  &_nextPrev{
-      padding-top: 15px;
-      margin-left: 30px;
-      
-      button:disabled,
-      button[disabled]{
-        opacity: 0.5;
-      }
+  &__navigation {
+    display: inline-block;
+    font-weight: normal;
+    text-align: center;
+    white-space: nowrap;
+    vertical-align: middle;
+    background-color: #ffffff;
+    background-image: none;
+    -ms-touch-action: manipulation;
+    touch-action: manipulation;
+    -webkit-user-select: none;
+    -moz-user-select: none;
+    -ms-user-select: none;
+    user-select: none;
+    border: none;
+    cursor: pointer;
 
-      img {
-        height: 10px;
-        width: 10px;
-        opacity: 3.0;
-      }
+    &:focus,
+    &:active &:active:focus {
+      outline: 0;
+      background-image: none;
+    }
 
-      &_prev {
-        margin-right: 15px;
-        background-color: Transparent;
-        background-repeat: no-repeat;
-        border: none;
-        cursor: pointer;
-        overflow: hidden;
-        outline: none;
-      }
-      &_next {
-        margin-left: 15px;
-        background-color: Transparent;
-        background-repeat: no-repeat;
-        border: none;
-        cursor: pointer;
-        overflow: hidden;
-        outline: none;
-        -webkit-transform:rotate(180deg);
-        -moz-transform: rotate(180deg);
-        -ms-transform: rotate(180deg);
-        -o-transform: rotate(180deg);
-        transform: rotate(180deg);
-      }
+    &:disabled {
+      cursor: not-allowed;
+      opacity: 0.5;
+    }
+
+    &___left {
+      height: 1.25em;
+      margin-right: 0.5em;
+    }
+
+    &___right {
+      transform: rotateY(180deg);
+      height: 1.25em;
+      margin: 0 0.5em;
+    }
   }
 }
