--- src/components/request/Pagination/Pagination.scss
@@ -1,80 +1,0 @@
-.pagination-container {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  padding: 1.5em 1em;
-  font-size: 1em;
-  font-family: 'Poppins', sans-serif;
-  color: #555 !important;
-  background: #ffffff;
-}
-
-.pagination-section {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  &__text {
-    margin: 0 1.25em;
-
-    span {
-      color: #555 !important;
-    }
-  }
-
-  &__select {
-    width: 75px;
-    display: inline-block;
-    padding: 2px 6px 2px 0;
-    box-sizing: border-box;
-    line-height: 1.42857143;
-    background-color: #ffffff;
-    background-image: none;
-    border: none;
-    outline: none;
-    -webkit-box-shadow: inset 0 -2px rgba(0, 0, 0, 0.075);
-    box-shadow: inset 0 -2px rgba(0, 0, 0, 0.075);
-    -webkit-transition: border-color ease-in-out 0.15s, -webkit-box-shadow ease-in-out 0.15s;
-    -o-transition: border-color ease-in-out 0.15s, box-shadow ease-in-out 0.15s;
-    transition: border-color ease-in-out 0.15s, box-shadow ease-in-out 0.15s;
-  }
-
-  &__navigation {
-    display: inline-block;
-    font-weight: normal;
-    text-align: center;
-    white-space: nowrap;
-    vertical-align: middle;
-    background-color: #ffffff;
-    background-image: none;
-    -ms-touch-action: manipulation;
-    touch-action: manipulation;
-    -webkit-user-select: none;
-    -moz-user-select: none;
-    -ms-user-select: none;
-    user-select: none;
-    border: none;
-    cursor: pointer;
-
-    &:focus,
-    &:active &:active:focus {
-      outline: 0;
-      background-image: none;
-    }
-
-    &:disabled {
-      cursor: not-allowed;
-    }
-
-    &___left {
-      height: 1.25em;
-      margin-right: 0.5em;
-    }
-
-    &___right {
-      transform: rotateY(180deg);
-      height: 1.25em;
-      margin: 0 0.5em;
-    }
-  }
-}
