+++ src/pages/Request/Request.scss
@@ -1,3 +1,65 @@
 .request-page {
   padding: 1.25em;
 }
+.big-container{
+  position: fixed;  
+  width: 100%;  
+  height: 100%;  
+  left: 0;  
+  right: 0;  
+  bottom: 0;  
+  margin: auto;  
+  background: #ECECEC;
+  display: block;
+  font-family: 'Poppins', sans-serif;
+  font-size: 13px;
+
+}
+
+.sub-container{
+  display: block;
+  width: 79%;
+  height: 100%;
+  margin-left: 19%;
+  margin-top: 7%;
+
+  .mainButton{
+    width: 100%;
+    background: white;
+    margin-left: 0%;
+
+    .buton{
+      width: 20%;
+      margin-left: 1%;
+      margin-bottom: 10px;
+    }  
+  }
+  
+}
+.line{
+  opacity: 0.3;
+  border: 1px solid #C4C4C4;
+}
+.emptyDiv{
+  width: 100%;
+  height: 20px;
+  background: #FFFFFF;
+}
+.search{
+  width: 100%;
+  height: 30px;
+  background: #FFFFFF;
+  &-box {    
+    position: absolute;
+    display: flex;
+    justify-content: space-between;
+    width: 150px;
+    height: 25px;
+    left: 80%;
+    margin-top: -15px;
+  }
+  .page-title {
+    padding: 5px 0 0 20px;
+    display: inline-block;
+  }
+}
