--- src/pages/TripRequest/TripRequest.scss
@@ -1,57 +1,0 @@
-.big-container{
-  position: fixed;  
-  width: 100%;  
-  height: 100%;  
-  left: 0;  
-  right: 0;  
-  bottom: 0;  
-  margin: auto;  
-  background: #ECECEC;
-  display: block;
-  font-family: 'Poppins', sans-serif;
-  font-size: 13px;
-
-}
-
-.sub-container{
-  display: block;
-  background: white;
-  width: 79%;
-  height: 100%;
-  margin-left: 19%;
-  margin-top: 7%;
-
-  .mainButton{
-    width: 100%;
-    margin-left: 1%;
-
-    .buton{
-      width: 20%;
-      margin-left: 0%;
-      margin-bottom: 10px;
-    }  
-  }
-  
-}
-.line{
-  opacity: 0.3;
-  border: 1px solid #C4C4C4;
-}
-.search{
-  width: 100%;
-  height: 30px;
-  &-box {    
-    position: absolute;
-    display: flex;
-    justify-content: space-between;
-    background: #FFFFFF;
-    width: 150px;
-    height: 25px;
-    left: 80%;
-    margin-top: -15px;
-  }
-  .page-title {
-    padding: 5px 0 0 20px;
-    display: inline-block;
-  }
-}
