+++ src/store/actions/types.js
@@ -1,4 +1,3 @@
 export const LOGIN_FAILURE = 'LOGIN_FAILURE';
 export const SUCCESS_LOGIN = 'SUCCESS_LOGIN';
 export const SUCCESS_GET_USER = 'SUCCESS_GET_USER';
-export const SUCCESS_GET_REQUESTS = 'SUCCESS_GET_REQUESTS';
